% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{fibR}
\alias{fibR}
\title{Gives the nth value in a fibonacci sequenct.}
\usage{
fibR(x)
}
\arguments{
\item{x}{}
}
\description{
\code{meanR()} Runs the same function as dcpp_fibonacci but in the R language.
}
\examples{

fibR(1:10)
}

