% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{slopeR}
\alias{slopeR}
\title{Gives the slope between two vectors}
\usage{
slopeR(x, y)
}
\arguments{
\item{x}{}

\item{y}{}
}
\description{
\code{slopeR()} Runs the same function as dcpp_slope but in the R language.
}

